Don't assume that RRTR uses an SSRC that's one of ours
